doubles arent supported correctly on mips/mipsel platform
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@7687 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/services/networking/madwifi.c
@@ -1150,7 +1150,7 @@ set_rate (char *dev)
     }
   else
     {
-      double ratef = atof (r) * 1000.0;
+      float ratef = atof (r) * 1000.0;
       int integerrate = (int) ratef;
       char set[32];
       sprintf (set, "%d", integerrate);
